Move FtbTeamsTeamProvider to compat
