+++ mod/src/main/java/org/shsts/tinactory/integration/tech/TechManagers.java
@@ -6,7 +6,6 @@ import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.saveddata.SavedData;
-import net.neoforged.fml.ModList;
 import org.shsts.tinactory.api.tech.ITeamProfile;
 import org.shsts.tinactory.api.tech.ITeamProvider;
 import org.shsts.tinactory.core.tech.TechInitPacket;
@@ -19,7 +18,6 @@ import org.shsts.tinycorelib.api.network.PacketDirection;
 import java.util.Optional;
 
 import static org.shsts.tinactory.Tinactory.REGISTRATE;
-import static org.shsts.tinactory.TinactoryConfig.CONFIG;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
@@ -78,23 +76,12 @@ public final class TechManagers {
         client = new ClientTechManager();
     }
 
-    public static void installTeamProvider() {
-        var provider = switch (CONFIG.teamProvider.get()) {
-            case SINGLE_PLAYER -> new SinglePlayerTeamProvider();
-            case FTB_TEAMS -> createFtbTeamsProvider();
-        };
+    public static void installTeamProvider(ITeamProvider provider) {
         server().installTeamProvider(provider);
     }
 
     public static void uninstallTeamProvider() {
         server().uninstallTeamProvider();
-    }
-
-    private static ITeamProvider createFtbTeamsProvider() {
-        if (!ModList.get().isLoaded("ftbteams")) {
-            throw new IllegalStateException("FTB Teams provider requires the ftbteams mod");
-        }
-        return new FtbTeamsTeamProvider(server(), CONFIG.ftbUnaffiliatedPlayerPolicy.get());
     }
 
     public static void loadSavedData(ServerLevel world) {
